+++ sdk/appconfiguration/azure-appconfiguration-provider/tests/test_provider.py
@@ -6,7 +6,7 @@
 from azure.appconfiguration.provider import SettingSelector, AzureAppConfigurationKeyVaultOptions
 from devtools_testutils import recorded_by_proxy
 from preparers import app_config_decorator
-from testcase import AppConfigTestCase
+from testcase import AppConfigTestCase, has_feature_flag
 import datetime
 from unittest.mock import patch
 
@@ -23,12 +23,14 @@ class TestAppConfigurationProvider(AppConfigTestCase):
     @app_config_decorator
     def test_provider_creation(self, appconfiguration_connection_string, appconfiguration_keyvault_secret_url):
         client = self.create_client(
-            appconfiguration_connection_string, keyvault_secret_url=appconfiguration_keyvault_secret_url
+            appconfiguration_connection_string,
+            keyvault_secret_url=appconfiguration_keyvault_secret_url,
+            feature_flag_enabled=True,
         )
         assert client["message"] == "hi"
         assert client["my_json"]["key"] == "value"
-        assert "FeatureManagementFeatureFlags" in client
-        assert "Alpha" in client["FeatureManagementFeatureFlags"]
+        assert "FeatureManagement" in client
+        assert has_feature_flag(client, "Alpha")
 
     # method: provider_trim_prefixes
     @recorded_by_proxy
@@ -39,13 +41,14 @@ class TestAppConfigurationProvider(AppConfigTestCase):
             appconfiguration_connection_string,
             trim_prefixes=trimmed,
             keyvault_secret_url=appconfiguration_keyvault_secret_url,
+            feature_flag_enabled=True,
         )
         assert client["message"] == "hi"
         assert client["my_json"]["key"] == "value"
         assert client["trimmed"] == "key"
         assert "test.trimmed" not in client
-        assert "FeatureManagementFeatureFlags" in client
-        assert "Alpha" in client["FeatureManagementFeatureFlags"]
+        assert "FeatureManagement" in client
+        assert has_feature_flag(client, "Alpha")
 
     # method: provider_selectors
     @recorded_by_proxy
@@ -59,7 +62,7 @@ class TestAppConfigurationProvider(AppConfigTestCase):
         )
         assert client["message"] == "test"
         assert "test.trimmed" not in client
-        assert "FeatureManagementFeatureFlags" not in client
+        assert "FeatureManagement" not in client
 
     # method: provider_selectors
     @recorded_by_proxy