+++ sdk/keyvault/azure-keyvault-keys/tests/test_crypto_client_async.py
@@ -37,9 +37,7 @@ NO_GET = Permissions(keys=[p.value for p in KeyPermissions if p.value != "get"])
 
 all_api_versions = get_decorator(is_async=True)
 only_hsm = get_decorator(only_hsm=True, is_async=True)
-only_vault_7_4_plus = get_decorator(
-    only_vault=True, is_async=True, api_versions=[ApiVersion.V7_4, ApiVersion.V7_5_PREVIEW_1]
-)
+only_vault_7_4_plus = get_decorator(only_vault=True, is_async=True, api_versions=[ApiVersion.V7_4, ApiVersion.V7_5])
 no_get = get_decorator(is_async=True, permissions=NO_GET)